+++ front/src/Pages/CustomPage/OptionPage/index.jsx
@@ -7,11 +7,10 @@ import OptionCatalogue from "./OptionCatalogue";
 import RadioGroup from "@Components/Common/RadioGroup";
 import OptionTagLabel from "@Components/Custom/OptionTagLabel";
 import { useUserData } from "context/UserDataContext";
-import { postUserData } from "apis/custom";
 
 const tabData = ["추가 옵션", "기본 포함 옵션"];
 const OptionPage = () => {
-  const { totalData, 유저데이터저장 } = useUserData();
+  const { totalData, 유저데이터저장, formatAndPost } = useUserData();
   const handleSelectOption = (option) => {
     if (totalData["옵션"].map((option) => option.id).includes(option.id)) {
       유저데이터저장(
@@ -41,21 +40,6 @@ const OptionPage = () => {
   };
 
   const move = useButtonNavigation();
-  // 커스텀 정보 서버 전송
-  const formatAndPost = () => {
-    const { 트림, 엔진, 바디, 구동방식, 외장, 내장, 옵션 } = totalData;
-    const formattedData = {
-      trimId: 트림.id,
-      componentIds: [엔진.id, 바디.id, 구동방식.id],
-      interiorId: 내장.id,
-      exteriorId: 외장.id,
-      optionGroupIds: 옵션.map((option) => option.id),
-    };
-
-    postUserData(formattedData).then((data) => {
-      move(`/result/${data.code}`);
-    });
-  };
 
   return (
     <Wrapper>
